Simplify applied policy telemetry
